feat: new value system + refactoring
For now store values/locals as enums instead of raw bytes. The memory/performance overhead is not important as of now.

+++ examples/basic/main.rs
@@ -12,7 +12,7 @@ fn main() -> ExitCode {
 
     let wat = r#"
     (module
-        (func $add_one (param $x i32) (param $y i32) (result i32) (result i32) (local $ununsed_local f32)
+        (func $add_one (param $x i32) (param $y i32) (result i32) (result i32) (local $ununsed_local i32)
             local.get $y
             i32.const 1
             i32.add
